Proxy the realtime compiler server instead of extracting
This allows us to directly use the default file without having to keep a modified version up to date. The proxy script loads the Phar which could in theory make it slower, but the implementation is lighter, codewise. The `getPharPath` is also changed to return the latest build when running from source, providing some level of testing compatibility.