Remove development dependencies when building release
Due to the way that distributions for Packagist needs to be configured, according to the Laravel Zero docs (https://laravel-zero.com/docs/build-a-standalone-application#distribute-via-packagist), all dependencies need to be dev-dependencies.

This presents a problem. We don't want to include the actual development dependencies in the release. My best solution as of now is to simply remove them before building the release version. Sadly, this means we cannot cache the Composer packages for the job anymore.

Fixes https://github.com/hydephp/cli/issues/26