Remove realtime compiler server proxy caching
This causes issues when using different Hyde binaries in different locations. Just wasted half an hour finding out that the reason my code didn't change was because the server proxy was generated with the global version and thus used that archive's code, instead of the development build in my parent directory.

This change makes to that the proxy is regenerated each time the server starts, which should not have any real performance impact, especially since it is just when the serve command is run.